+++ examples/typescript/http-server-pool/express-hybrid/src/types.ts
@@ -1,7 +1,6 @@
-import type { Worker } from 'node:worker_threads'
-import type { PoolOptions } from 'poolifier'
+import type { ThreadPoolOptions } from 'poolifier'
 
-export interface ClusterWorkerData extends PoolOptions<Worker> {
+export interface ClusterWorkerData extends ThreadPoolOptions {
   port: number
   workerFile: string
   minWorkers?: number