+++ src/index.ts
@@ -1,6 +1,7 @@
 export type { AbstractPool } from './pools/abstract-pool'
 export { DynamicClusterPool } from './pools/cluster/dynamic'
 export { FixedClusterPool } from './pools/cluster/fixed'
+export type { ClusterPoolOptions } from './pools/cluster/fixed'
 export { PoolEvents, PoolTypes } from './pools/pool'
 export type {
   IPool,
@@ -34,6 +35,7 @@ export {
 } from './pools/selection-strategies/selection-strategies-types'
 export type {
   IWorkerChoiceStrategy,
+  InternalWorkerChoiceStrategyOptions,
   Measurement,
   MeasurementOptions,
   MeasurementStatisticsRequirements,
@@ -45,6 +47,7 @@ export type {
 export type { WorkerChoiceStrategyContext } from './pools/selection-strategies/worker-choice-strategy-context'
 export { DynamicThreadPool } from './pools/thread/dynamic'
 export { FixedThreadPool } from './pools/thread/fixed'
+export type { ThreadPoolOptions } from './pools/thread/fixed'
 export type { AbstractWorker } from './worker/abstract-worker'
 export { ClusterWorker } from './worker/cluster-worker'
 export { ThreadWorker } from './worker/thread-worker'